requirements.txt
Flask
google-generativeai
markdown
beautifulsoup4
